reorganize flipbook files
