+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/FlipbookViewer/FlipbookViewer.js
@@ -1,78 +1,16 @@
 import React, { useEffect, useRef, useState } from 'react'
 import { Box } from 'grommet'
-import asyncStates from '@zooniverse/async-states'
 import PropTypes from 'prop-types'
-import { withStores } from '@helpers'
 
 import locationValidator from '../../helpers/locationValidator'
-import useSubjectImage, {
-  placeholder
-} from '../SingleImageViewer/hooks/useSubjectImage'
+import { placeholder } from '../SingleImageViewer/hooks/useSubjectImage'
 
 import SingleImageViewer from '../SingleImageViewer/SingleImageViewer'
 import SVGImage from '../SVGComponents/SVGImage'
 import SVGPanZoom from '../SVGComponents/SVGPanZoom'
 import FlipbookControls from './components'
 
-function storeMapper(store) {
-  const { frame: defaultFrame } = store.subjectViewer
-
-  return {
-    defaultFrame
-  }
-}
-
-function FlipbookViewer({
-  defaultFrame = 0,
-  loadingState = asyncStates.initialized,
-  onError = () => true,
-  onReady = () => true,
-  subject
-}) {
-  console.log(defaultFrame, loadingState)
-  /** This initializes an image element from the subject's defaultFrame src url.
-   * We do this so the SVGPanZoom has dimensions of the subject image.
-   * We're assuming all frames in one subject have the same dimensions. */
-  const defaultFrameUrl = subject ? Object.values(subject.locations[defaultFrame])[0] : null
-  const { img, error } = useSubjectImage(window.Image, defaultFrameUrl)
-  const { naturalHeight, naturalWidth, src: defaultFrameSrc } = img
-
-  if (error) {
-    console.error(error)
-    onError(error)
-  }
-
-  if (loadingState === asyncStates.error || !subject?.locations) {
-    return <div>Something went wrong.</div>
-  }
-
-  return (
-    <FlipbookImage
-      defaultFrame={defaultFrame}
-      defaultFrameSrc={defaultFrameSrc}
-      naturalHeight={naturalHeight}
-      naturalWidth={naturalWidth}
-      onError={onError}
-      onReady={onReady}
-      subject={subject}
-    />
-  )
-}
-
-FlipbookViewer.propTypes = {
-  defaultFrame: PropTypes.number,
-  loadingState: PropTypes.string,
-  onError: PropTypes.func,
-  onReady: PropTypes.func,
-  subject: PropTypes.shape({
-    locations: PropTypes.arrayOf(locationValidator)
-  }).isRequired
-}
-
-export default withStores(FlipbookViewer, storeMapper)
-export { FlipbookViewer }
-
-export const FlipbookImage = ({
+const FlipbookViewer = ({
   defaultFrame,
   defaultFrameSrc,
   onReady,
@@ -93,7 +31,8 @@ export const FlipbookImage = ({
   useEffect(() => {
     const svgImage = subjectImage?.current
     if (svgImage && defaultFrameSrc !== placeholder.src) {
-      const { width: clientWidth, height: clientHeight } = svgImage.getBoundingClientRect()
+      const { width: clientWidth, height: clientHeight } =
+        svgImage.getBoundingClientRect()
       const target = { clientHeight, clientWidth, naturalHeight, naturalWidth }
       onReady({ target })
     }
@@ -132,3 +71,16 @@ export const FlipbookImage = ({
     </Box>
   )
 }
+
+FlipbookViewer.propTypes = {
+  defaultFrame: PropTypes.number,
+  defaultFrameSrc: PropTypes.string,
+  onReady: PropTypes.func,
+  naturalHeight: PropTypes.number,
+  naturalWidth: PropTypes.number,
+  subject: PropTypes.shape({
+    locations: PropTypes.arrayOf(locationValidator)
+  }).isRequired
+}
+
+export default FlipbookViewer
