+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/FlipbookViewer/index.js
@@ -1,1 +1,1 @@
-export { default } from './FlipbookViewer'
+export { default } from './FlipbookViewerContainer'
